requirements.txt
pandas==2.2.1
matplotlib==3.8.3
fpdf2==2.7.8
openpyxl==3.1.2
